+++ .ralph/rust-bindings-plan.state.json
@@ -1,0 +1,13 @@
+{
+  "name": "rust-bindings-plan",
+  "taskFile": ".ralph/rust-bindings-plan.md",
+  "iteration": 8,
+  "maxIterations": 50,
+  "itemsPerIteration": 3,
+  "reflectEvery": 10,
+  "reflectInstructions": "REFLECTION CHECKPOINT\n\nPause and reflect on your progress:\n1. What has been accomplished so far?\n2. What's working well?\n3. What's not working or blocking progress?\n4. Should the approach be adjusted?\n5. What are the next priorities?\n\nUpdate the task file with your reflection, then continue working.",
+  "active": true,
+  "status": "active",
+  "startedAt": "2026-05-13T06:32:01.135Z",
+  "lastReflectionAt": 0
+}
