+++ bindings/rust/crates/maplibre-native/src/map.rs
@@ -11,7 +11,9 @@ use maplibre_native_sys as sys;
 use crate::custom_geometry::CanonicalTileId;
 use crate::custom_geometry::CustomGeometrySourceState;
 use crate::events::MapId;
+use crate::geometry::GeometryNativeExt;
 use crate::handle::{ThreadAffineNativeHandle, closed_handle_error, out_handle};
+use crate::json::JsonValueNativeExt;
 use crate::render::{
     MetalBorrowedTextureDescriptor, MetalOwnedTextureDescriptor, MetalSurfaceDescriptor,
     OwnedTextureDescriptor, RenderSessionHandle, VulkanBorrowedTextureDescriptor,
